basic completed

+++ src/controllers/cart.controllers.ts
@@ -4,12 +4,15 @@ const { AppError, sendResponse, catchAsync } = utilsHelper
 import Cart from '~/models/database/Cart'
 import Product from '~/models/database/Product'
 import { updateProduct } from './product.controllers'
+import User from '~/models/database/User'
 
 export const createCart = catchAsync(async (req: Request, res: Response, next: NextFunction) => {
   //Get data from request
   const { product_id, quantity } = req.body
   const user_id = req.userId
   //Validation
+  const user = await User.findOne({ _id: user_id, isDeleted: false })
+  if (!user) throw new AppError(400, 'User not found', 'Get User Error')
   const checkProduct = await Product.findById(product_id)
   if (!checkProduct) throw new AppError(401, 'Product not exist', 'Create cart error')
   const checkCartExists = await Cart.findOne({ product_id, user_id })
@@ -22,9 +25,11 @@ export const createCart = catchAsync(async (req: Request, res: Response, next: N
 
 export const getAllProductsInCart = catchAsync(async (req: Request, res: Response, next: NextFunction) => {
   //Get data from request
-  const userId = req.userId
+  const user_id = req.userId
   //validation
-  const carts = await Cart.find({ user_id: userId }, '-user_id -_id').populate({
+  const user = await User.findOne({ _id: user_id, isDeleted: false })
+  if (!user) throw new AppError(400, 'User not found', 'Get User Error')
+  const carts = await Cart.find({ user_id }, '-user_id -_id').populate({
     path: 'product_id',
     model: 'Product',
     select: 'name price imageURL'
@@ -44,6 +49,9 @@ export const updateQuantityInCart = catchAsync(async (req: Request, res: Respons
   //Get data from request
   const { quantity, product_id } = req.body
   const userId = req.userId
+  //Validation
+  const user = await User.findOne({ _id: userId, isDeleted: false })
+  if (!user) throw new AppError(400, 'User not found', 'Get User Error')
   //Process
   if (quantity == 0) {
     await Cart.findOneAndDelete({ user_id: userId, product_id: product_id })
